[fix] change coordinate back to double

+++ Backend/Repository/RoasterRepository.cs
@@ -45,7 +45,7 @@ public class RoasterRepository(DataContext context) : BaseRepository<Roaster>(co
     }
 
     NetTopologySuite.Geometries.Point? searchPoint = null;
-    float? distanceInMeters = null;
+    double? distanceInMeters = null;
 
     // TODO: Should we add a validation to ensure all 3 exists
     if (roasterParams.Lat.HasValue &&
